Refactor GDrive & Sheets sources according to review comments
