+++ components/google_sheets/sources/new-row-added-shared-drive/new-row-added-shared-drive.mjs
@@ -10,10 +10,6 @@ export default {
   version: "0.0.1",
   dedupe: "unique",
   type: "source",
-  hooks: {
-    ...httpBase.hooks,
-    ...newRowAdded.hooks,
-  },
   props: {
     ...httpBase.props,
     ...newRowAdded.props,
