+++ components/google_sheets/sources/new-updates/new-updates.mjs
@@ -10,10 +10,6 @@ export default {
   description: "Emit new event each time a row or cell is updated in a spreadsheet. To use this source with a spreadsheet in a [Shared Drive](https://support.google.com/a/users/answer/9310351), use the **New Updates (Shared Drive, Instant)** source instead.",
   version: "0.0.26",
   dedupe: "unique",
-  hooks: {
-    ...httpBase.hooks,
-    ...newUpdates.hooks,
-  },
   props: {
     ...httpBase.props,
     ...newUpdates.props,
